feat: handle Stripe sessionId in an isolated component
